Move SetFocus call into the main process() routine
Move SetFocus call into the main process() routine
With SetFocus where it was in localProcess() the special libraries
(CheckBox, EditBox, etc..) will reset the status to NOT_EXECUTED.
Moving it to process() allows us to properly process SetFocus and
maintain proper status without having to change all the subclasses.
Also, found some code that was in the tadvm CVS that was NOT in the
version pushed to GIT.  This was regarding the REFRESH of the
libComponent if it was found to be null.