Add support for AutoConfigureJSAFS to injectRuntimeAwarenessClasses
Modified some code to avoid duplication in Lists of classes.  Added
support to interrogate classes being checked for the RuntimeAware
interface to also be checked for the AutoConfigureJSAFS annotation.
This allows a test to include and exclude additional runtime packages
that might be in different packages from other projects and
dependencies.